Refactoring and fixes. * Modified Listener to use a single Addr field instead of pk/port. * Modified PortManager to have an internal local PK. * Split server.go into server.go and server_conn.go * Fixed behavior and tests.
