+++ client.go
@@ -57,7 +57,6 @@ type Client struct {
 
 	pm *PortManager
 
-	// accept map[uint16]chan *transport
 	done chan struct{}
 	once sync.Once
 }
@@ -70,10 +69,8 @@ func NewClient(pk cipher.PubKey, sk cipher.SecKey, dc disc.APIClient, opts ...Cl
 		sk:    sk,
 		dc:    dc,
 		conns: make(map[cipher.PubKey]*ClientConn),
-		pm:    newPortManager(),
-		// accept: make(chan *transport, AcceptBufferSize),
-		// accept: make(map[uint16]chan *transport),
-		done: make(chan struct{}),
+		pm:    newPortManager(pk),
+		done:  make(chan struct{}),
 	}
 	for _, opt := range opts {
 		if err := opt(c); err != nil {
@@ -103,7 +100,7 @@ func (c *Client) updateDiscEntry(ctx context.Context) error {
 
 func (c *Client) setConn(ctx context.Context, conn *ClientConn) {
 	c.mx.Lock()
-	c.conns[conn.remoteSrv] = conn
+	c.conns[conn.srvPK] = conn
 	if err := c.updateDiscEntry(ctx); err != nil {
 		c.log.WithError(err).Warn("updateEntry: failed")
 	}
@@ -213,7 +210,7 @@ func (c *Client) findOrConnectToServer(ctx context.Context, srvPK cipher.PubKey)
 		return nil, err
 	}
 
-	conn := NewClientConn(c.log, nc, c.pk, srvPK, c.pm)
+	conn := NewClientConn(c.log, c.pm, nc, c.pk, srvPK)
 	if err := conn.readOK(); err != nil {
 		return nil, err
 	}
@@ -244,7 +241,7 @@ func (c *Client) findOrConnectToServer(ctx context.Context, srvPK cipher.PubKey)
 
 // Listen creates a listener on a given port, adds it to port manager and returns the listener.
 func (c *Client) Listen(port uint16) (*Listener, error) {
-	l, ok := c.pm.NewListener(c.pk, port)
+	l, ok := c.pm.NewListener(port)
 	if !ok {
 		return nil, errors.New("port is busy")
 	}
